split file tree tests
