+++ tests/integration/file-import-export/add-files.spec.ts
@@ -9,25 +9,11 @@ import * as XLSX from 'xlsx';
 import { createFile } from '../../utils';
 import { test as dataViewTest } from '../fixtures/data-view';
 import { test as dbExplorerTest } from '../fixtures/db-explorer';
-import { FILE_SYSTEM_TREE, test as filePickerTest } from '../fixtures/file-picker';
-import { test as fileSystemExplorerTest } from '../fixtures/file-system-explorer';
+import { test as filePickerTest } from '../fixtures/file-picker';
 import { test as baseTest } from '../fixtures/page';
-import { test as scriptEditor } from '../fixtures/script-editor';
 import { test as spotlightTest } from '../fixtures/spotlight';
-import { test as storageTest } from '../fixtures/storage';
-import { test as testTmpTest } from '../fixtures/test-tmp';
 
-const test = mergeTests(
-  baseTest,
-  storageTest,
-  filePickerTest,
-  testTmpTest,
-  fileSystemExplorerTest,
-  dataViewTest,
-  spotlightTest,
-  scriptEditor,
-  dbExplorerTest,
-);
+const test = mergeTests(baseTest, filePickerTest, dataViewTest, spotlightTest, dbExplorerTest);
 
 test('should add csv files and folders', async ({
   addFileButton,
@@ -299,127 +285,4 @@ test('should handle duckdb files with reserved names correctly', async ({
     // Confirm the deletion
     await assertDBExplorerItems([]);
   }
-});
-
-test('should create file tree structure and verify persistence after reload', async ({
-  filePicker,
-  clickFileByName,
-  assertFileExplorerItems,
-  page,
-  reloadPage,
-  renameFileInExplorer,
-  assertDBExplorerItems,
-  renameDBInExplorer,
-  setupFileSystem,
-}) => {
-  await page.goto('/');
-
-  expect(filePicker).toBeDefined();
-
-  // Create files and directories
-  await setupFileSystem(FILE_SYSTEM_TREE);
-
-  // 5. Check the file tree structure
-  const rootStructure = ['dir-a', 'a', 'a_1 (a)', 'parquet-test', 'xlsx-test'];
-
-  await assertFileExplorerItems(rootStructure);
-
-  const firstLevelStructure = [
-    'dir-a',
-    'dir-b',
-    'a_4 (a)',
-    'a_5 (a)',
-    'a',
-    'a_1 (a)',
-    'parquet-test',
-    'xlsx-test',
-  ];
-  const secondLevelStructure = [
-    'dir-a',
-    'dir-b',
-    'a_2 (a)',
-    'a_3 (a)',
-    'a_4 (a)',
-    'a_5 (a)',
-    'a',
-    'a_1 (a)',
-    'parquet-test',
-    'xlsx-test',
-  ];
-
-  const checkFileTreeStructure = async () => {
-    // First, check the root level
-    await assertFileExplorerItems(rootStructure);
-    // Click on the 'dir-a' folder to open its contents
-    await clickFileByName('dir-a');
-    // Check the contents of the 'dir-a' folder (including files and the 'dir-b' folder)
-    await assertFileExplorerItems(firstLevelStructure);
-    // Click on the 'dir-b' folder to open its contents
-    await clickFileByName('dir-b');
-    // Check the contents of the 'dir-b' folder
-    await assertFileExplorerItems(secondLevelStructure);
-  };
-  await checkFileTreeStructure();
-
-  // 6. Reload the page and re-check persistence
-  await reloadPage();
-
-  // Repeat checks after reload
-  await checkFileTreeStructure();
-
-  // 7. Check the DB explorer
-  await page.getByTestId('navbar-show-databases-button').click();
-  await assertDBExplorerItems(['testdb', 'main', 'test_view']);
-
-  // 8. Rename files and check persistence
-  await reloadPage();
-
-  // Rename files
-  await renameFileInExplorer({
-    oldName: 'a',
-    newName: 'a_renamed',
-    expectedNameInExplorer: 'a_renamed (a)',
-  });
-  await renameFileInExplorer({
-    oldName: 'a_1 (a)',
-    newName: 'a_1_renamed',
-    expectedNameInExplorer: 'a_1_renamed (a)',
-  });
-  await renameFileInExplorer({
-    oldName: 'parquet-test',
-    newName: 'parquet_renamed',
-    expectedNameInExplorer: 'parquet_renamed (parquet-test)',
-  });
-  await renameFileInExplorer({
-    oldName: 'xlsx-test',
-    newName: 'xlsx_renamed',
-    expectedNameInExplorer: 'xlsx_renamed (xlsx-test)',
-  });
-
-  // Check the file tree structure after renaming
-  const rootWithRenamedFiles = [
-    'dir-a',
-    'a_renamed (a)',
-    'a_1_renamed (a)',
-    'parquet_renamed (parquet-test)',
-    'xlsx_renamed (xlsx-test)',
-  ];
-  await assertFileExplorerItems(rootWithRenamedFiles);
-
-  // 9. Switch to Databases tab and rename the DuckDB database
-  await page.getByTestId('navbar-show-databases-button').click();
-
-  await renameDBInExplorer({
-    oldName: 'testdb',
-    newName: 'testdb_renamed',
-    expectedNameInExplorer: 'testdb_renamed (testdb)',
-  });
-
-  // Check that the renamed DB appears
-  await assertDBExplorerItems(['testdb_renamed (testdb)', 'main', 'test_view']);
-
-  // 10. Reload the page and check persistence
-  await reloadPage();
-  await page.getByTestId('navbar-show-databases-button').click();
-  await assertDBExplorerItems(['testdb_renamed (testdb)', 'main', 'test_view']);
 });
