+++ tests/integration/fixtures/file-picker.ts
@@ -12,76 +12,6 @@ import { test as testTmpTest } from './test-tmp';
 import { createFile, parsePath } from '../../utils';
 import { FileSystemNode } from '../models';
 
-export const FILE_SYSTEM_TREE: FileSystemNode[] = [
-  {
-    type: 'file',
-    ext: 'csv',
-    content: 'col\ndata1',
-    name: 'a',
-  },
-  {
-    type: 'file',
-    ext: 'json',
-    content: '{"col": "data2"}',
-    name: 'a',
-  },
-  {
-    type: 'file',
-    ext: 'xlsx',
-    content: '[{"col": "dataXLSX1"}]',
-    name: 'xlsx-test',
-  },
-  {
-    type: 'file',
-    ext: 'parquet',
-    content: "SELECT 'data3' AS col;",
-    name: 'parquet-test',
-  },
-  {
-    type: 'file',
-    ext: 'duckdb',
-    content: 'CREATE VIEW test_view AS SELECT 123 AS value;',
-    name: 'testdb',
-  },
-  {
-    type: 'dir',
-    name: 'dir-a',
-    children: [
-      {
-        type: 'file',
-        ext: 'csv',
-        content: 'col\ndataA1',
-        name: 'a',
-      },
-      {
-        type: 'file',
-        ext: 'json',
-        content: '{"col": "dataA2"}',
-        name: 'a',
-      },
-
-      {
-        type: 'dir',
-        name: 'dir-b',
-        children: [
-          {
-            type: 'file',
-            ext: 'csv',
-            content: 'col\ndataB1',
-            name: 'a',
-          },
-          {
-            type: 'file',
-            ext: 'json',
-            content: '{"col": "dataB2"}',
-            name: 'a',
-          },
-        ],
-      },
-    ],
-  },
-];
-
 interface FilePicker {
   // Playwright does not support it (https://github.com/microsoft/playwright/issues/8850).
   // So we have to patch the files and directory pickers Browser API.
